🔧 chore(.env): add POCKETBASE_URL environment variable ✨ feat(LoginForm.tsx): add login functionality The login function now sends a request to the server to authenticate the user. If the authentication is successful, the user is redirected to the home page. If there is an error, the error message is displayed to the user.
✨ feat(RegisterForm.tsx): add registration functionality
The register function now sends a request to the server to register the user. If the registration is successful, the user is redirected to the login page. If there is an error, the error message is displayed to the user.

🔧 chore(NavMenu.tsx): add getUser and logout functions
The getUser function retrieves the user model from the local storage. The logout function clears the authentication store.

🔧 chore(SiteHeader.tsx): comment out login link and uncomment NavDropdownMenu
The login link is commented out and the NavDropdown